# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
servers:
  - url: http://localhost:6185/api/v1
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /stats:
    get:
      tags: [Stats]
      summary: Get runtime statistics
      operationId: getStats
      x-astrbot-scope: system
      parameters:
        - name: offset_sec
          in: query
          schema:
            type: integer
            default: 86400
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /stats/provider-tokens:
    get:
      tags: [Stats]
      summary: Get provider token usage statistics
      operationId: getProviderTokenStats
      x-astrbot-scope: system
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /stats/first-notice:
    get:
      tags: [Stats]
      summary: Get first-run dashboard notice content
      operationId: getFirstNotice
      x-astrbot-scope: system
      parameters:
        - name: locale
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /stats/ghproxy/test:
    post:
      tags: [Stats]
      summary: Test a GitHub proxy endpoint
      operationId: testGhproxyConnection
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GhproxyTestRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /changelogs/{version}:
    get:
      tags: [Stats]
      summary: Get changelog content for a version
      operationId: getChangelog
      x-astrbot-scope: system
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "404":
          $ref: "#/components/responses/Error"
...
  /stats/storage/cleanup:
    post:
      tags: [Stats]
      summary: Clean storage
      operationId: cleanupStorage
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /backups:
    get:
      tags: [Backups]
      summary: List backups
      operationId: listBackups
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Backups]
      summary: Export a backup
      operationId: createBackup
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupExportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/upload:
    post:
      tags: [Backups]
      summary: Upload a backup file
      operationId: uploadBackup
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/BackupUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/upload/init:
    post:
      tags: [Backups]
      summary: Initialize chunked backup upload
      operationId: initBackupUpload
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupUploadInitRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/upload/chunk:
    post:
      tags: [Backups]
      summary: Upload a backup chunk
      operationId: uploadBackupChunk
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/BackupChunkUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/upload/complete:
    post:
      tags: [Backups]
      summary: Complete chunked backup upload
      operationId: completeBackupUpload
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupUploadSessionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/upload/abort:
    post:
      tags: [Backups]
      summary: Abort chunked backup upload
      operationId: abortBackupUpload
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupUploadSessionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/tasks/{task_id}:
    get:
      tags: [Backups]
      summary: Get backup task progress
      operationId: getBackupProgress
      x-astrbot-scope: system
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/{filename}:
    get:
      tags: [Backups]
      summary: Download a backup
      operationId: downloadBackup
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/Filename"
      responses:
        "200":
          description: Backup archive
    patch:
      tags: [Backups]
      summary: Rename a backup
      operationId: renameBackup
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/Filename"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupRenameRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Backups]
      summary: Delete a backup
      operationId: deleteBackup
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/Filename"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/{filename}/check:
    post:
      tags: [Backups]
      summary: Check a backup before import
      operationId: checkBackup
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/Filename"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /backups/{filename}/import:
    post:
      tags: [Backups]
      summary: Import a backup
      operationId: importBackup
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/Filename"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupImportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /updates/releases:
    get:
      tags: [Updates]
      summary: List releases
      operationId: listReleases
      x-astrbot-scope: system
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [core, dashboard]
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /updates/core:
    post:
      tags: [Updates]
      summary: Update AstrBot core
      operationId: updateCore
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /updates/dashboard:
    post:
      tags: [Updates]
      summary: Update dashboard assets
      operationId: updateDashboard
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /updates/progress/{task_id}:
    get:
      tags: [Updates]
      summary: Get update progress
      operationId: getUpdateProgress
      x-astrbot-scope: system
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /pip/install:
    post:
      tags: [Updates]
      summary: Install a Python package
      operationId: installPipPackage
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipInstallRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /migrations:
    post:
      tags: [Updates]
      summary: Run pending migrations
      operationId: runMigrations
      x-astrbot-scope: system
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrationRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /cron/jobs:
    get:
      tags: [Cron]
      summary: List cron jobs
      operationId: listCronJobs
      x-astrbot-scope: system
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Cron]
      summary: Create a cron job
      operationId: createCronJob
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CronJobRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /cron/jobs/{job_id}:
    patch:
      tags: [Cron]
      summary: Update a cron job
      operationId: updateCronJob
      x-astrbot-scope: system
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CronJobPatchRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Cron]
      summary: Delete a cron job
      operationId: deleteCronJob
      x-astrbot-scope: system
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /cron/jobs/{job_id}/run:
    post:
      tags: [Cron]
      summary: Run a cron job immediately
      operationId: runCronJob
      x-astrbot-scope: system
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /trace/settings:
    get:
      tags: [Logs]
      summary: Get trace settings
      operationId: getTraceSettings
      x-astrbot-scope: system
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Logs]
      summary: Update trace settings
      operationId: updateTraceSettings
      x-astrbot-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TraceSettingsRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /t2i/templates:
    get:
      tags: [T2I]
      summary: List text-to-image templates
      operationId: listT2iTemplates
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [T2I]
      summary: Create a text-to-image template
      operationId: createT2iTemplate
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/T2iTemplateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /t2i/templates/active:
    get:
      tags: [T2I]
      summary: Get active text-to-image template
      operationId: getActiveT2iTemplate
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [T2I]
      summary: Set active text-to-image template
      operationId: setActiveT2iTemplate
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /t2i/templates/{name}:
    get:
      tags: [T2I]
      summary: Get a text-to-image template
      operationId: getT2iTemplate
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [T2I]
      summary: Update a text-to-image template
      operationId: updateT2iTemplate
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/T2iTemplateContentRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [T2I]
      summary: Delete a text-to-image template
      operationId: deleteT2iTemplate
      x-astrbot-scope: config
      parameters:
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /subagents/config:
    get:
      tags: [Subagents]
      summary: Get subagent orchestrator configuration
      operationId: getSubagentConfig
      x-astrbot-scope: config
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Subagents]
      summary: Update subagent orchestrator configuration
      operationId: updateSubagentConfig
      x-astrbot-scope: config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    Error:
      description: Standard AstrBot error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  parameters:
...
    Filename:
      name: filename
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  schemas:
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    BackupExportRequest:
      type: object
      properties:
        include:
          type: array
          items:
            type: string
        exclude:
          type: array
          items:
            type: string
      additionalProperties: false
    BackupUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    BackupUploadInitRequest:
      type: object
      required: [filename, total_size]
      properties:
        filename:
          type: string
        total_size:
          type: integer
          minimum: 1
      additionalProperties: false
    BackupUploadSessionRequest:
      type: object
      required: [upload_id]
      properties:
        upload_id:
          type: string
      additionalProperties: false
    BackupChunkUploadRequest:
      type: object
      required: [upload_id, chunk_index, chunk]
      properties:
        upload_id:
          type: string
        chunk_index:
          type: integer
          minimum: 0
        chunk:
          type: string
          format: binary
    BackupRenameRequest:
      type: object
      required: [new_name]
      properties:
        new_name:
          type: string
      additionalProperties: false
    BackupImportRequest:
      type: object
      properties:
        confirmed:
          type: boolean
          default: true
      additionalProperties: false
    UpdateRequest:
      type: object
      properties:
        version:
          type: string
        proxy:
          type: string
        reboot:
          type: boolean
        progress_id:
          type: string
      additionalProperties: false
...
    GhproxyTestRequest:
      type: object
      required: [proxy_url]
      properties:
        proxy_url:
          type: string
      additionalProperties: false
...